+++ components/listings/show/ListingSlider/index.js
@@ -5,7 +5,7 @@ import FontAwesomeIcon from '@fortawesome/react-fontawesome'
 import faAngleRight from '@fortawesome/fontawesome-pro-regular/faAngleRight'
 import faAngleLeft from '@fortawesome/fontawesome-pro-regular/faAngleLeft'
 import faCube from '@fortawesome/fontawesome-free-solid/faCube'
-import CloseButton from 'components/shared/CloseButton'
+import CloseButton from 'components/shared/Common/Buttons/CloseButton'
 import ButtonIcon from 'components/shared/Common/Buttons'
 import {thumbnailUrl} from 'utils/image_url'
 import {downloadBlob} from 'utils/file-utils'
