+++ components/shared/Common/Buttons/CloseButton/index.js
