+++ components/shared/Common/Buttons/CloseButton/style.js
